w13 updated

+++ Desktop/КБТУ/Web/week13/hh_back/api/views/views.py
@@ -1,19 +1,22 @@
 from rest_framework import status
+from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from api.models import Company, Vacancy
-from api.serializers import CompanySerializer2, VacancySerializer
+from api.serializers import CompanySerializer, VacancySerializer
 
 
-class CompanyListAPIView(APIView):
-    def get(self, request):
+@api_view(['GET', 'POST'])
+def companies(request):
+    if request.method == 'GET':
         c = Company.objects.all()
-        serializer = CompanySerializer2(c, many=True)
+        serializer = CompanySerializer(c, many=True)
+
         return Response(serializer.data)
 
-    def post(self, request):
-        serializer = CompanySerializer2(data=request.data)
+    elif request.method == 'POST':
+        serializer = CompanySerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -22,47 +25,43 @@ class CompanyListAPIView(APIView):
                         status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-class CompanyDetailAPIView(APIView):
-    def get_object(self, company_id):
-        try:
-            return Company.objects.get(id=company_id)
-        except Company.DoesNotExist as e:
-            return Response({'error': str(e)})
+@api_view(['GET', 'PUT', 'DELETE'])
+def company(request, company_id):
+    try:
+        c = Company.objects.get(id=company_id)
+    except Company.DoesNotExist as e:
+        return Response({'error': str(e)})
 
-    def get(self, request, company_id):
-        c = self.get_object(company_id)
-        serializer = CompanySerializer2(c)
+    if request.method == 'GET':
+        serializer = CompanySerializer(c)
         return Response(serializer.data)
 
-    def put(self, request, company_id):
-        c = self.get_object(company_id)
-        serializer = CompanySerializer2(instance=c, data=request.data)
+    elif request.method == 'PUT':
+        serializer = CompanySerializer(instance=c, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
 
         return Response({'error': serializer.errors})
 
-    def delete(self, request, company_id):
-        c = self.get_object(company_id)
+    elif request.method == 'DELETE':
         c.delete()
-
         return Response({'deleted': True})
 
 
-class CompanyWithVacanciesListAPView(APIView):
-    def get(self, request, company_id):
-        v = Vacancy.objects.filter(id=company_id)
-        serializer = VacancySerializer(v, many=True)
+@api_view(["GET"])
+def company_vacancies(request, company_id):
+    try:
+        c = Company.objects.get(id=company_id)
+        serializer = VacancySerializer(
+            Vacancy.objects.filter(company=c),
+            many=True
+        )
         return Response(serializer.data)
-
-    def post(self, request):
-        serializer = VacancySerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response({'error': serializer.errors},
-                        status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+    except Company.DoesNotExist as e:
+        return Response({'error': str(e)})
+    except Vacancy.DoesNotExist as e:
+        return Response({'error': str(e)})
 
 
 class VacancyListAPIView(APIView):
@@ -76,7 +75,6 @@ class VacancyListAPIView(APIView):
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-
         return Response({'error': serializer.errors},
                         status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
@@ -99,11 +97,18 @@ class VacancyDetailAPIView(APIView):
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-
         return Response({'error': serializer.errors})
 
     def delete(self, request, vacancy_id):
         v = self.get_object(vacancy_id)
         v.delete()
-
         return Response({'deleted': True})
+
+
+class VacancyTopTenAPIView(APIView):
+    def get(self, request):
+        serializer = VacancySerializer(
+            Vacancy.objects.order_by('-salary', 'name')[:10],
+            many=True
+        )
+        return Response(serializer.data)
